requirements.txt
scipy==1.7.1
numpy==1.21.2
matplotlib==3.5.0
opencv_python==4.5.4.60
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
Pillow==8.4.0
h5py==2.10.0
